+++ src/pages/OSpy/components/Replayer/index.tsx
@@ -5,20 +5,15 @@ import { Button, Flex, Space } from 'antd';
 import { useTranslation } from 'react-i18next';
 import { ArrowLeftOutlined } from '@ant-design/icons';
 import { useSize } from 'ahooks';
-import { Link, useLocation, useNavigate } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { SelectLogButton } from '@/components/SelectLogButton';
+import { useUrlParam } from '@/utils/useUrlParam';
 
 export const Replayer = () => {
   const { t } = useTranslation();
   const size = useSize(document.body);
   const navigate = useNavigate();
-
-  // DON'T USE useSearch OR URLSearchParams, it will decode the url value automatically
-  const { search } = useLocation();
-  const replayUrl = useMemo(() => {
-    const url = search.split('?url=')?.[1];
-    return url || '';
-  }, [search]);
+  const replayUrl = useUrlParam();
 
   const backSlot = useMemo(() => {
     return (
